Refactoring function value loader and invoker
