+++ src/WebJobs.Script/Description/DiagnosticSeverityExtensions.cs
@@ -1,0 +1,35 @@
+﻿// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the MIT License. See License.txt in the project root for license information.
+
+using System;
+using System.Diagnostics;
+using Microsoft.CodeAnalysis;
+
+namespace Microsoft.Azure.WebJobs.Script.Description
+{
+    [CLSCompliant(false)]
+    public static class DiagnosticSeverityExtensions
+    {
+        public static TraceLevel ToTraceLevel(this DiagnosticSeverity severity)
+        {
+            var level = TraceLevel.Off;
+            switch (severity)
+            {
+                case DiagnosticSeverity.Hidden:
+                    level = TraceLevel.Verbose;
+                    break;
+                case DiagnosticSeverity.Info:
+                    level = TraceLevel.Info;
+                    break;
+                case DiagnosticSeverity.Warning:
+                    level = TraceLevel.Warning;
+                    break;
+                case DiagnosticSeverity.Error:
+                    level = TraceLevel.Error;
+                    break;
+            }
+
+            return level;
+        }
+    }
+}
